Trait polkadot_omni_node_lib::CliConfig

pub trait CliConfig {
    // Required methods
    fn impl_version() -> String;
    fn author() -> String;
    fn support_url() -> String;
    fn copyright_start_year() -> u16;

    // Provided method
    fn description(executable_name: String) -> String { ... }
}

Trait that can be used to customize some of the customer-facing info related to the node binary that is being built using this library.

The related info is shown to the customer as part of logs or help messages. It does not impact functionality.

Required Methods

fn impl_version() -> String

The version of the resulting node binary.

fn author() -> String

The author of the resulting node binary.

fn support_url() -> String

The support URL for the resulting node binary.

fn copyright_start_year() -> u16

The starting copyright year of the resulting node binary.

Provided Methods

fn description(executable_name: String) -> String

The description of the resulting node binary.

Object Safety

This trait is not object safe.